% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_bioc_dependencies.R
\name{install_bioc_dependencies}
\alias{install_bioc_dependencies}
\title{Install Bioconductor dependencies}
\usage{
install_bioc_dependencies(bioc.args = list(), force = FALSE)
}
\arguments{
\item{bioc.args}{list containing further arguments to be passed
down to `BiocManager::install()`.}

\item{force}{logical: reinstall already-installed packages?}
}
\value{
No return value, called for side effects.
}
\description{
This function installs the latest versions of all Bioconductor packages
required for the report generation, namely:
\itemize{
   \item AnnotationDbi
   \item KEGGREST
   \item GO.db
}
}
\details{
It is essential that 
these Bioconductor packages be installed for CALANGO to work properly.
It uses `BiocManager::install()` for installing Bioconductor
packages. Further arguments to this function are passed as a list.
}
\examples{
\dontrun{
  install_bioc_dependencies()
}

}
