% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_calanguize_genomes.R
\name{retrieve_calanguize_genomes}
\alias{retrieve_calanguize_genomes}
\title{Retrieve calanguize_genomes script from the Github repository}
\usage{
retrieve_calanguize_genomes(
  target.dir,
  method = "auto",
  unzip = getOption("unzip")
)
}
\arguments{
\item{target.dir}{path to the folder where the files will be saved (
accepts relative and absolute paths)}

\item{method}{Method to be used for downloading files. Current download
methods are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto": see _Details_ and _Note_ in the
documentation of \code{utils::download.file()}.}

\item{unzip}{The unzip method to be used. See the documentation of
\code{utils::unzip()} for details.}
}
\value{
No return value, called for side effects (see Description).
}
\description{
This script downloads the *calanguize_genomes.pl* Perl script from the 
repository, together with associated README instructions for using the 
script, managing dependencies, etc. It will extract the data into a folder 
containing everything that is needed for preparing data for using 
CALANGO.
}
\details{
If the `target.dir` provided does not exist it is created
(recursively) by the function.
}
\examples{
\dontrun{
  CALANGO::retrieve_calanguize_genomes(target.dir = "./data")
}

}
